Bug fixes, new trackers and set updates
Make prowl description work correctly
Make move torrent files locally after adding to remove Deluge work
correctly
—Note— I have hacked up my local copy of plugins/modify/set_fields.py
to accept a list instead of dict (so the item rendering is done in
order) and to accept an optional “render: no” flag. You will need to
reformat any “set:” block to match the formatting in the Flexget wiki
page for the set plugin to make it work correctly for your copy.